feat: Start implementing on tablet show player in a DialogFragment. In PlayerFragment extends change from Fragment to DialogFragment.
Because PlayerActivity uses a fragment transaction, it is already compatible with a DialogFragment.
In other words, the app still works after this commit with no other changes. However tablet is not yet displaying DialogFragment as a floating "popup"
Reference
Udacity SpotifyStreamer Stage 2 rubric for tablet - "The Now Playing controls are displayed in a DialogFragment"